pyproject.toml
[tool.poetry]
name = "spotinfer"
version = "0.1.0"
description = "A simple Python package for inference tasks"
authors = ["Tymon Żarski <tymon.zarski@gmail.com>"]
readme = "README.md"
packages = [{include = "spotinfer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
datacrunch = "^1.14.0"
typer = "^0.19.2"
rich = "^14.1.0"
python-dotenv = "^1.0.0"
inquirer = "^3.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
spotinfer = "spotinfer.main:cli"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=4.3.0,<5.0.0"
pytest = ">=8.4.2,<9.0.0"
pytest-cov = ">=4.0.0,<5.0.0"
pytest-mock = "^3.11.1"
build = ">=1.0.0,<2.0.0"
twine = ">=4.0.0,<5.0.0"
tox = ">=4.0.0,<5.0.0"
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^6.0.1"
autopep8 = "^2.3.2"

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".tox",
    ".eggs",
    "*.egg",
]

[tool.autopep8]
max_line_length = 88
aggressive = 1

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true
